feat(FeedScheduler): 피드 삭제 로직 개선 및 프로젝트 연계 삭제 기능 추가
- FeedScheduler 클래스에 프로젝트 연계 삭제 조건을 추가하여 피드 삭제 로직을 개선함.
- 주요 변경 사항:
  1. deleteFeeds 메서드에서 삭제 가능한 피드를 결정하기 전에 관련 프로젝트의 완료 상태 및 완료 후 경과 시간을 체크하는 로직을 구현.
  2. canDeleteFeedBasedOnProjects 메서드를 추가하여 모든 관련 프로젝트가 완료 상태이고 지정된 기간(2개월)이 경과한 경우에만 피드를 삭제할 수 있도록 함.
  3. deleteProjects 메서드를 추가하여 피드와 연계된 프로젝트를 일괄 삭제하는 기능 구현.
  4. 기존 delete 메서드를 deleteFeed로 명확화하여 리팩토링.

- 이 변경은 시스템의 데이터 정합성을 높이고, 사용자가 프로젝트 완료 후 일정 기간이 지난 피드만 삭제할 수 있도록 하여 데이터 관리의 정확성을 향상시킴.
- 스케줄러의 효율성과 프로젝트 및 피드 관리의 일관성을 높이는 중요한 기능 추가.

+++ src/main/java/sidepair/service/scheduler/FeedScheduler.java
@@ -7,7 +7,9 @@ import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 import sidepair.domain.feed.Feed;
 import sidepair.domain.feed.FeedStatus;
+import sidepair.domain.project.Project;
 import sidepair.persistence.feed.FeedRepository;
+import sidepair.persistence.project.ProjectRepository;
 import sidepair.service.aop.ExceptionConvert;
 
 @Component
@@ -15,16 +17,36 @@ import sidepair.service.aop.ExceptionConvert;
 @RequiredArgsConstructor
 @ExceptionConvert
 public class FeedScheduler {
+    private static final int DELETE_AFTER_MONTH = 2;
 
     private final FeedRepository feedRepository;
+    private final ProjectRepository projectRepository;
 
     @Scheduled(cron = "0 0 4 * * *")
     public void deleteFeeds() {
         final FeedStatus status = FeedStatus.DELETED;
         final List<Feed> deletedStatusFeeds = feedRepository.findWithFeedContentByStatus(status);
         for (final Feed feed : deletedStatusFeeds) {
+            delete(feed);
+        }
+    }
+
+    private void delete(final Feed feed) {
+        final List<Project> projects = projectRepository.findByFeed(feed);
+        final boolean canDelete = canDeleteFeedBasedOnProjects(projects);
+        if (canDelete) {
+            deleteProjects(projects);
             deleteFeed(feed);
         }
+    }
+
+    private boolean canDeleteFeedBasedOnProjects(final List<Project> projects) {
+        return projects.stream()
+                .allMatch(project -> project.isCompleted() && project.isCompletedAfterMonths(DELETE_AFTER_MONTH));
+    }
+
+    private void deleteProjects(final List<Project> projects) {
+        projectRepository.deleteAll(projects);
     }
 
     private void deleteFeed(final Feed feed) {
